Add interface to regularize in the Open Closed Principle.
