+++ src/main/java/br/pegz/tutorials/rightcourt/serve/PlayService.java
@@ -6,7 +6,7 @@ import br.pegz.tutorials.rightcourt.persistence.enums.Side;
 import br.pegz.tutorials.rightcourt.persistence.enums.Speed;
 import br.pegz.tutorials.rightcourt.score.ScoreNotifierService;
 import br.pegz.tutorials.rightcourt.serve.exception.PointException;
-import br.pegz.tutorials.rightcourt.serve.resource.CourtResource;
+import br.pegz.tutorials.rightcourt.serve.resource.RestCourtResource;
 import com.google.common.collect.Lists;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -20,24 +20,24 @@ import java.util.Random;
 @Service
 public class PlayService {
 
-    private CourtResource courtResource;
+    private RestCourtResource restCourtResource;
     private Random random = new SecureRandom();
     private ScoreNotifierService scoreNotifierService;
 
-    public PlayService(CourtResource courtResource, ScoreNotifierService scoreNotifierService) {
-        this.courtResource = courtResource;
+    public PlayService(RestCourtResource restCourtResource, ScoreNotifierService scoreNotifierService) {
+        this.restCourtResource = restCourtResource;
         this.scoreNotifierService = scoreNotifierService;
     }
 
     public Flux<Play> serve() throws PointException {
         List<Play> plays = Lists.newArrayList(getServePlay());
-        Play incomingPlay = courtResource.sendPlayToOtherSide(getServePlay());
+        Play incomingPlay = restCourtResource.sendPlayToOtherSide(getServePlay());
         while (canContinuePlay(incomingPlay)) {
             log.info("Received play: {}", incomingPlay);
             plays.add(incomingPlay);
             Play myPlay = handlePlay(incomingPlay);
             plays.add(myPlay);
-            incomingPlay = courtResource.sendPlayToOtherSide(myPlay);
+            incomingPlay = restCourtResource.sendPlayToOtherSide(myPlay);
         }
         notifyPoint(incomingPlay);
         return Flux.fromStream(plays.stream());
